.claude/worktrees/quizzical-mayer/requirements.txt
fastapi
uvicorn
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-prometheus
psutil
redis[hiredis]
python-dotenv
mlflow